pyproject.toml
[tool.poetry]
name = "proj"
version = "0.1.0"
description = "Python project template using Poetry as package manager."
authors = ["kengz <kengzwl@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "^1.3.2"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
autopep8 = "^2.3.1"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--cov=proj",
    "--cov-report=html",
    "--cov-report=term",
]
filterwarnings = "ignore"

[tool.coverage.report]
include_namespace_packages = true
